Create journey-tests.yaml
We believe this will trigger the UAT tests after a build - and turn red if the tests fail. 

This isn't ideal, since the artifact will already be published to ECR.  We're talking to CDP team about whether there's a better way to do this that genuinely blocks making a releasable artifact on journey test failure.