DSFAAP-449: use standard format for guard clauses
FWIW, I don't think this version is as readable (and definitely not as
concise) as the it replaces, but team majority finds the braces version
more readable. And it keeps sonarcloud happy.

I don't think the complaint from sonarcloud itself is legitimate - since
the first stanza exits the function, there's a strong indication that
the next line is not included in the if condition.