BAU: validation & state into reusable model
The suggested model is:

`fromState` & `toState` handles translating to and from a raw payload -
suitable for showing on forms.

This lets us use (e.g) simple strings for cphNumbers or onOffRadio, and
use complex data types (e.g. for address) in the state.

At the same time, we can provide `validate` methods against the raw
payload, which means that errors map on to the structure that the UI
expects.