DSFAAP-583: short-circuit page generation on invalid answer
If a user has submitted an invalid answer at any point in the journey,
then there is no well defined output for the nextPage function on a
QuestionPage.

e.g. if you were to not have selected an input to 'Email' or 'Post' for
your licence delivery preference, then that page wouldn't be able to
calculate a next page.