DSFAAP-630: UNTESTED - use no-store on GET pages
This forces the browser (at least, in latest version of firefox) to
reload pages such as the /check-answers pages and the /task-list pages
to show the most recent application state when the user uses the back
button.

Right now, when you press the back button, you have to refresh before
those pages reflect the most recent application state, which can lead to
some confusion.