BAU: add build badges to enhance visibility
The default action page shows a list of breakages - but it doesn't make
it clear (without some extra eyeball work) which jobs have returned to a
good state.

As a result, it's quite easy to miss - in particular - the extended browserstack test
failures.

This makes the things that we want to be green all the time obvious -
the publish & the extended tests, which both run off the main branch.