BAU: remove memory-leaking logger we don't yet use
Running unit tests, it was warning of an event emitter leak - and
slowing down substantially over time.

Using `node --trace-warnings node_modules/.bin/jest -i`, it identified
the createLogger() function call within each of the subclasses.

This is because properties are initialised on class instantiation, so
every time we were creating an instance for our classes under test, we
were creating a new logger.

I'm removing this for now - to prevent any issues in development - since
we're not actively using this logger.

When we need to re-introduce it, we can find a more pino-friendly way to
do it.