BAU: remove duplicate metrics
In the CDP platform, we have an nginx cluster in front of the services
that gives us logs with messages that look like:

`[response] get /origin/type-of-origin 200 (8ms)`

There are also common fields parsed out - e.g referrer, path, etc

This is already being rendered into a count graph in grafana (based on a
lucene query against those logs).  We could make version of the graph
that splits that by method and/or url too (with CDP's help), since both
are parsed as separate fields from the nginx log.

We don't need to additionally log this manually from the service itself.