DSFAAP-790: disable textarea copy-paste detection
We've got a huge amount of change happening, and lowering cognative load
for us as a team is a large part of staying productive.

There *is* a lot of duplication between this and the input type, but
blending them into one class is going to cause quite a lot of
harder-to-follow logic (e.g. whether an email type is appropriate, the
format of the viewModel).

Once we've done with the R2 scope, then we can revisit these
similarities & see if there is any other factored-out step we want to
include.

But for now, keeping our base AnswerModel structure recognisable and
clear is far more important than small amounts of duplication.