BAU: allow routing to depend on whole app state
In Release 2, we have some pages that show up depending on *both* the
answer to the current question and prior application state.

To facilitate this, routing decisions need to know about both:
- the immediate answer given
- the historical answers given

For instance, after the origin type question, the user is routed two a
different CPH number & Address page depending on whether you're going on
the premises or off the premises.