DSFAAP-853: support config factory for radio button
This allows us to:

1. change the *option* text, without changing the value
2. change the *order* of options, even when options are the same

The first one is necessary for this ticket - where the concept of
'other' premise has an updated label (and yet we're implementing this in
a way that preserves the current implementation that's accessible in
live).

The second one will be necessary for future tickets (specifically
destination premises type), where the order for each option has been
user test specifically & helps users answer the question.