BAU: simplify test login management
In the application, there are 4 cookies:

- defra-id-bell
- userSession
- session
- crumb

The login-specific cookies are defra-id-bell and userSession - with
userSession being the one that tells the application that the user is
successfully signed in.

Session and (to a lesser extent) crumb are the cookies that manage
application state and user journey aside from authorisation.

e.g. where are you in the journey, and whether your page submission is
believed to be legitimate.

This PR deletes only the application state cookies, keeping the auth
ones in place.

Also, by doing this we avoid global state, which opens up to (e.g.)
having multiple test accounts with different roles in future, without
tests conflicting with each other.