BAU: add basic page templating & make template improvements
Our workflow means that we need to add a basic page for each
nextPage destination that we define.

As a result, we want:
- an easy way to produce a basic page with the right values
- an easy way to replace that with a fully fledged implementation
  that aligns with what we've already defined

re-using the same questions gives us this quick way to do it, and
forcing the overwrite of the index.js file when we're using the
question-page template makes it easy to turn that into the full
implementation.

Other smaller improvements:

- adding 'number' as a question type
- replacing missed cases when a class is referred to directly with
  its corresponding template value or TEMPLATE-TODO if a template
  value does not readily exist