BAU: await next page when entering text details
In the current implementation, we initiate the request to wait for
a new page, but we don't wait for it to be complete before we go on
to the next step.

For slightly slower browsers (Safari), this is causing the back button to be
hit on the **current** page - not the next page - resulting
in the browser window closing.