BAU: remove old feature flag references
We previously supported **both** a feature flag to switch sending
to case management, and a query parameter to force it (so we could smoke test
the integration in live without having to turn the feature on for all users).

But there is no feature flag in the config anymore, nor any specifically supported
query params.

So this test code can be removed.

(also, by convention, our type imports are at the top underneath the code imports.
So making a minor correction there too)