BAU: select the autocomplete input, not select box
Currently, it will select whatever has the #disinfectant API, and await
on that before it puts any input in.

However, this could be the original page select box, rather than the javascript-offered
text box.

This creates a race condition that seems to be causing test failures - as
wdio tries to treat a select box as an input box.